refactor Rename to fixtures to utilities
